Log the firmware version when it's queried. (dm)

+++ b2g/ui/src/org/nbp/b2g/ui/FirmwareVersion.java
@@ -1,8 +1,8 @@
 package org.nbp.b2g.ui;
 
 public abstract class FirmwareVersion {
-  public static native byte getMajor ();
-  public static native byte getMinor ();
+  public static native int getMajor ();
+  public static native int getMinor ();
 
   private FirmwareVersion () {
   }
